update: renamed ContractResultLogEntry V2 into main version
