+++ src/components/contract/ContractResultLogs.vue
@@ -40,7 +40,7 @@
     </template>
     <template v-slot:content>
       <template v-for="l in nbLogDisplayed" :key="l">
-        <ContractResultLogEntryV2 :log="logs[logCursor + l - 1]" :block-number="blockNumber" :transaction-hash="transactionHash"/>
+        <ContractResultLogEntry :log="logs[logCursor + l - 1]" :block-number="blockNumber" :transaction-hash="transactionHash"/>
         <hr class="h-card-separator" style="height: 1px; background: grey"/>
       </template>
 
@@ -73,8 +73,8 @@
 
 import {computed, defineComponent, onMounted, PropType, Ref, ref, watch} from "vue";
 import DashboardCard from "@/components/DashboardCard.vue";
-import ContractResultLogEntryV2 from "@/components/contract/ContractResultLogEntryV2.vue";
-import {ContractLog, ContractResultLog} from "@/schemas/HederaSchemas";
+import ContractResultLogEntry from "@/components/contract/ContractResultLogEntry.vue";
+import {ContractLog} from "@/schemas/HederaSchemas";
 import {AppStorage} from "@/AppStorage";
 
 const DEFAULT_PAGE_SIZE = 3
@@ -84,7 +84,7 @@ export default defineComponent({
   name: "ContractResultLogs",
 
   components: {
-    ContractResultLogEntryV2,
+    ContractResultLogEntry,
     DashboardCard
   },
